requirements.txt
aiogram==2.12.1
Flask==1.1.2
Flask-Login==0.5.0
Flask-RESTful==0.3.8
Flask-WTF==0.14.3
Pillow==8.1.0
requests==2.25.1
SQLAlchemy==1.3.23
SQLAlchemy-serializer==1.3.4.4
WTForms==2.3.3